style: update overview tab component
