+++ frontend/src/components/molecules/AgentCard.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import React from 'react';
 
 import { useRouter } from 'next/navigation';
